refactor(notebook): extract shell process adapter
